+++ src/sys/mem/mod.rs
@@ -9,9 +9,9 @@ pub use paging::{
 };
 pub use phys::{phys_addr, PhysBuf};
 
+use crate::sys::boot::MemoryMap;
 use crate::sys::pic;
 
-use bootloader::bootinfo::{BootInfo, MemoryMap};
 use core::sync::atomic::{AtomicUsize, Ordering};
 use spin::Once;
 use x86_64::structures::paging::{OffsetPageTable, Translate};
@@ -21,10 +21,9 @@ use x86_64::{PhysAddr, VirtAddr};
 static mut MAPPER: Once<OffsetPageTable<'static>> = Once::new();
 
 static PHYS_MEM_OFFSET: Once<u64> = Once::new();
-static MEMORY_MAP: Once<&MemoryMap> = Once::new();
 static MEMORY_SIZE: AtomicUsize = AtomicUsize::new(0);
 
-pub fn init(boot_info: &'static BootInfo) {
+pub fn init(memory_map: &MemoryMap, offset: u64) {
     // Keep the timer interrupt to have accurate boot time measurement but mask
     // the keyboard interrupt that would create a panic if a key is pressed
     // during memory allocation otherwise.
@@ -32,10 +31,9 @@ pub fn init(boot_info: &'static BootInfo) {
 
     let mut memory_size = 0;
     let mut last_end_addr = 0;
-    for region in boot_info.memory_map.iter() {
-        let start_addr = region.range.start_addr();
-        let end_addr = region.range.end_addr();
-        let size = end_addr - start_addr;
+    for region in memory_map.iter() {
+        let start_addr = region.addr;
+        let end_addr = region.addr + region.size;
         let hole = start_addr - last_end_addr;
         if hole > 0 {
             log!(
@@ -48,9 +46,9 @@ pub fn init(boot_info: &'static BootInfo) {
         }
         log!(
             "MEM [{:#016X}-{:#016X}] {:?}", // "({} KB)"
-            start_addr, end_addr - 1, region.region_type //, size >> 10
+            start_addr, end_addr - 1, region.kind //, size >> 10
         );
-        memory_size += size as usize;
+        memory_size += region.size as usize;
         last_end_addr = end_addr;
     }
 
@@ -66,13 +64,12 @@ pub fn init(boot_info: &'static BootInfo) {
     unsafe {
         MAPPER.call_once(|| OffsetPageTable::new(
             paging::active_page_table(),
-            VirtAddr::new(boot_info.physical_memory_offset),
+            VirtAddr::new(offset),
         ))
     };
 
-    PHYS_MEM_OFFSET.call_once(|| boot_info.physical_memory_offset);
-    MEMORY_MAP.call_once(|| &boot_info.memory_map);
-    bitmap::init_frame_allocator(&boot_info.memory_map);
+    PHYS_MEM_OFFSET.call_once(|| offset);
+    bitmap::init_frame_allocator(memory_map);
     heap::init_heap().expect("heap initialization failed");
 
     pic::unmask(pic::KBD_IRQ);